backend/requirements.txt
fastapi 
uvicorn[standard] 
python-dotenv 
pydantic 
anthropic 
openai 
motor 
pymongo 
dnspython 
certifi 
